fix(repo-user): all tests working now due to fixing little bugs in repo

+++ apps/repo-user/src/main/java/com/epita/repo_user/service/UserService.java
@@ -64,7 +64,7 @@ public class UserService implements Logger {
         AuthService.generateToken(userModel.getId().toString(), userModel.getUsername()));
   }
 
-  @Transactional
+  // @Transactional
   public UserEntity updateUser(ObjectId id, ModifyUserRequest request) {
     if (request == null) throw RepoUserErrorCode.INVALID_USER_DATA.createError("request");
 
@@ -91,7 +91,7 @@ public class UserService implements Logger {
 
     userModel.setUpdatedAt(LocalDateTime.now());
 
-    userRepository.persist(userModel);
+    userRepository.update(userModel);
     logger().info("New User model registered: '{}'", userModel.getId());
 
     redisPublisher.publish(
@@ -137,7 +137,7 @@ public class UserService implements Logger {
     throw RepoUserErrorCode.USER_NOT_FOUND.createError();
   }
 
-  @Transactional
+  // @Transactional
   public void deleteUser(ObjectId id) {
     UserModel userModel =
         userRepository
@@ -145,11 +145,11 @@ public class UserService implements Logger {
             .orElseThrow(() -> RepoUserErrorCode.USER_NOT_FOUND.createError(id.toString()));
     userModel.setDeleted(true);
     UserAggregate userAggregate = userModelToUserAggregate.convertNotNull(userModel);
-    // userRepository.deleteById(id); // TODO: delete this ?
+    userRepository.deleteById(id);
     redisPublisher.publish(userAggregateChannel, userAggregate);
   }
 
-  @Transactional
+  // @Transactional
   public UserEntity uploadProfileImage(ObjectId userId, UploadImageRequest request) {
     String objectKey = "user/" + userId + "/image/" + UUID.randomUUID() + ".jpeg";
 
@@ -160,9 +160,12 @@ public class UserService implements Logger {
     try {
       s3Service.uploadFile(
           objectKey, request.getFile(), request.getFile().available()); // TODO: fix size
-      s3Service.deleteFile(userModel.getProfileImage());
+      if (userModel.getProfileImage() != null)
+      {
+        s3Service.deleteFile(userModel.getProfileImage());
+      }
       userModel.setProfileImage(objectKey);
-      userRepository.persist(userModel);
+      userRepository.update(userModel);
       redisPublisher.publish(
           userAggregateChannel, userModelToUserAggregate.convertNotNull(userModel));
       return userModelToUserEntityConverter.convertNotNull(userModel);
